# Conflicts:
#	common/src/main/java/org/geogebra/common/euclidian/EuclidianController.java
#	common/src/main/java/org/geogebra/common/kernel/geos/GeoInline.java

+++ retex/renderer-base/src/main/java/com/himamis/retex/renderer/share/platform/graphics/stubs/GraphicsStub.java
@@ -86,7 +86,7 @@ public class GraphicsStub implements Graphics2DInterface {
 	}
 
 	@Override
-	public void fillRect(int x, int y, int width, int height) {
+	public void fillRect(double x, double y, double width, double height) {
 		// stub
 	}
